// types/characters.ts
export type Character = {
  name: string;
  height: number;
  mass: number;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  films: string;
  species: string[];
  vehicles: string[];
  starships: string[];
  created: Date;
  edited: Date;
  url: string;
};

export type GetCharactersResults = {
  count: number;
  next: string | null;
  previous: string | null;
  results: Character[];
};